+++ src/main/java/performative/tasks/Deadline.java
@@ -1,18 +1,15 @@
-package performative;
+package performative.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Event extends Task {
-    private LocalDateTime start;
-    private LocalDateTime end;
+public class Deadline extends Task {
+    private LocalDateTime by;
 
-
-    public Event(String description, String start, String end) {
+    public Deadline(String description, String by) throws DateTimeParseException {
         super(description);
-        this.start = parseDateTime(start);
-        this.end = parseDateTime(end);
+        this.by = parseDateTime(by);
     }
 
     private LocalDateTime parseDateTime(String dateTimeString) throws DateTimeParseException {
@@ -27,12 +24,13 @@ public class Event extends Task {
 
     @Override
     public String toSaveFormat() {
+        // Save in original YYYY-MM-DD HHMM format
         DateTimeFormatter saveFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm");
-        return "performative.Event; " + (super.getStatus() ? "Complete" : "Incomplete") + "; " + super.getDescription() + "; " + start.format(saveFormatter) + "; " + end.format(saveFormatter);
+        return "performative.tasks.Deadline; " + (super.getStatus() ? "Complete" : "Incomplete") + "; " + super.getDescription() + "; " + by.format(saveFormatter);
     }
 
     @Override
     public String toString() {
-        return "[E]" + super.toString() + " (from: " + formatDateTime(this.start) + ", to: " + formatDateTime(this.end) + ")";
+        return "[D]" + super.toString() + " (by: " + formatDateTime(this.by) + ")";
     }
 }
