+++ src/main/java/performative/tasks/Todo.java
@@ -1,4 +1,4 @@
-package performative;
+package performative.tasks;
 
 public class Todo extends Task {
     public Todo (String description) {
@@ -7,7 +7,7 @@ public class Todo extends Task {
 
     @Override
     public String toSaveFormat() {
-        return "performative.Todo; " + (super.getStatus() ? "Complete" : "Incomplete") + "; " + super.getDescription();
+        return "performative.tasks.Todo; " + (super.getStatus() ? "Complete" : "Incomplete") + "; " + super.getDescription();
     }
     @Override
     public String toString() {
