+++ src/main/java/performative/ui/Ui.java
@@ -1,6 +1,6 @@
 package performative.ui;
 
-import performative.Task;
+import performative.tasks.Task;
 
 import java.util.ArrayList;
